12-NVIDIA NIM PDF RAG/requirements.txt
openai==1.95.1
langchain==0.3.26
langchain-core==0.3.67
langchain_community==0.3.26
langchain_nvidia_ai_endpoints==0.3.10
faiss-cpu==1.11.0
pypdf==5.7.0 
python-dotenv
streamlit==1.46.1